WIP: add run from spinup task
